Improve schema file organization
