+++ src/api/namespaces/schema.ts
@@ -54,53 +54,6 @@ export const NamespaceLogListSchema = z.object({
 
 export const NamespaceDeletedSchema = z.null();
 
-// Regex for input validation. This isn't perfect but should be good enough for
-// a start. Matches git@hostname:path, where path isn't very restrictive.
-export const gitUrlSchema = z
-  .string()
-  .regex(/^([a-zA-Z0-9.\-_]+@[a-zA-Z0-9.\-_]+:[a-zA-Z0-9.\-_/]+)*$/, {
-    message: "format must be git@host:path when using SSH",
-  });
-
-// note: in the current API implementation, a mirror is created
-// by creating a namespace with the mirror object in the payload.
-
-export const MirrorPublicPostSchema = z.object({
-  url: z.string().url().nonempty(),
-  ref: z.string().nonempty(),
-});
-
-// When Token auth is used, token is submitted as "passphrase"
-export const MirrorTokenPostSchema = z.object({
-  url: z.string().url().nonempty(),
-  ref: z.string().nonempty(),
-  passphrase: z
-    .string()
-    .nonempty({ message: "Required when using token auth" }),
-});
-
-export const MirrorSshPostSchema = z.object({
-  url: gitUrlSchema.nonempty({
-    message: "format must be git@host:path when using SSH",
-  }),
-  ref: z.string().nonempty(),
-  passphrase: z.string().optional(),
-  privateKey: z.string().nonempty({ message: "Required when using SSH" }),
-  publicKey: z.string().nonempty({ message: "Required when using SSH" }),
-});
-
-export const MirrorKeepSSHKeysFormSchema = z.object({
-  url: gitUrlSchema.nonempty({
-    message: "format must be git@host:path when using SSH",
-  }),
-  ref: z.string().nonempty(),
-});
-
-export const MirrorPostSchema = MirrorPublicPostSchema.or(
-  MirrorTokenPostSchema
-).or(MirrorSshPostSchema);
-
 export type NamespaceListSchemaType = z.infer<typeof NamespaceListSchema>;
 export type NamespaceLogListSchemaType = z.infer<typeof NamespaceLogListSchema>;
 export type NamespaceLogSchemaType = z.infer<typeof NamespaceLogSchema>;
-export type MirrorPostSchemaType = z.infer<typeof MirrorPostSchema>;
